requirements.txt
altair==4.2.0
anyio==3.6.1
appnope==0.1.3
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.10.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fastapi==0.85.1
filelock==3.8.0
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.29
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.10.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
Levenshtein==0.20.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parrot==1.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==3.20.0
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pydantic==1.10.2
pydeck==0.8.0b4
pyfiglet==0.8.post1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-Levenshtein==0.20.5
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
rapidfuzz==2.11.1
regex==2022.9.13
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.2
scipy==1.9.2
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.5.1
starlette==0.20.4
streamlit==1.13.0
termcolor==2.0.1
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.22.2
typer==0.6.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
validators==0.20.0
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
zipp==3.10.0
